+++ admin_ordering/__init__.py
@@ -1,5 +1,2 @@
-from __future__ import absolute_import, unicode_literals
-
-
 VERSION = (0, 14, 1)
 __version__ = ".".join(map(str, VERSION))