+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Change log
 - Updated the bundled jQuery UI library to 1.12.1.
 - Added a system check which verifies that ``admin_ordering`` is added to
   ``INSTALLED_APPS``.
+- Added system checks for the ordering attribute of models.
 
 
 `0.14`_ (2021-03-29)